Move intents state from MainViewModel to IntentUtils
_intents/intents StateFlow and addAll/removeAll/clear are now owned by
IntentUtils (singleton), so intent state is co-located with intent
parsing logic. MainViewModel retains getAccount/showBunkerRequests/
onNewIntent and reads IntentUtils.intents.value directly. AccountScreen,
MainActivity, and SignerActivity updated to call IntentUtils directly.
The getIntentData duplicate-check no longer needs to reach back into the
activity to find the ViewModel.
