+++ app/src/main/java/com/greenart7c3/nostrsigner/MainActivity.kt
@@ -28,6 +28,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.rememberNavController
 import com.greenart7c3.nostrsigner.okhttp.HttpClientManager
 import com.greenart7c3.nostrsigner.service.BunkerRequestUtils
+import com.greenart7c3.nostrsigner.service.IntentUtils
 import com.greenart7c3.nostrsigner.ui.AccountScreen
 import com.greenart7c3.nostrsigner.ui.AccountStateViewModel
 import com.greenart7c3.nostrsigner.ui.CenterCircularProgressIndicator
@@ -188,7 +189,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     override fun onDestroy() {
-        mainViewModel.clear()
+        IntentUtils.clear()
 
         super.onDestroy()
     }
